+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ## Unreleased
 
 * 🐛 Make sure streams created with a different version of the polyfill do not pass the brand checks. ([#75](https://github.com/MattiasBuelens/web-streams-polyfill/issues/75), [#77](https://github.com/MattiasBuelens/web-streams-polyfill/pull/77))
+* 👓 Align with [spec version `c5ca883`](https://github.com/whatwg/streams/tree/c5ca8837c0b50ad1057351a7676824013c36538e/) ([#79](https://github.com/MattiasBuelens/web-streams-polyfill/pull/79))
 
 ## v3.0.3 (2020-04-09)